backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]==3.2.4
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.26.0
openai==1.12.0
groq==0.4.2
python-multipart==0.0.9
jinja2==3.1.3
python-docx==1.1.0
reportlab==4.1.0
apscheduler==3.10.4
